requirements.txt
Django>=4.0,<5.0
django-ninja
gunicorn
djangorestframework
djangorestframework-simplejwt
pandas
openpyxl